+++ Mario/Animation.h
@@ -10,7 +10,7 @@ using namespace std;
 
 class Animation
 {
-	ULONGLONG lastFrameTime;
+	DWORD lastFrameTime;
 	int defaultTime;
 	int currentFrame;
 	vector<LPANIMATION_FRAME> frames;
